Updates component identification with colors in LEGO set, updates AFD calculation
Turns out that some pieces had occupancy > 1, which should not be possible.
The reason was that some pieces occurred in a set in multiple colors, and so
were effectively double counted when grouping by `component_id` alone. Now,
it merges the original `component_id` with `color_id` to create a new unique
`component_id` for every piece.
For the AFD calculation, it now filters more reasonably.